Update 'next-3.x' with latest changes from 'main'

See merge request quiqqer/core!453

+++ admin/ajax/users/address/deleteAddressList.php
@@ -43,6 +43,27 @@ QUI::$Ajax->registerFunction(
             try {
                 $User = QUI::getUsers()->get($result[0]['userUuid']);
                 $Address = $User->getAddress($id);
+            } catch (QUI\Exception $Exception) {
+                QUI\System\Log::writeDebugException($Exception);
+
+                continue;
+            }
+
+            $StandardAddress = $User->getStandardAddress();
+
+            if (
+                $StandardAddress !== null
+                && $StandardAddress->getUUID() === $Address->getUUID()
+            ) {
+                throw new QUI\Exception(
+                    QUI::getLocale()->get(
+                        'quiqqer/core',
+                        'exception.users.address.delete.default'
+                    )
+                );
+            }
+
+            try {
                 $Address->delete();
             } catch (QUI\Exception $Exception) {
                 QUI\System\Log::writeDebugException($Exception);
